requirements.txt
pyyaml==6.0.1
python-dotenv==1.0.0
telethon==1.33.1